Hard reboot on project
Decided to rebuild the project from the ground up, this time using Visual Studio community. The goal from now on is to build all functionality of the program in c++ and use c#/.NET as the interface, later on this may be extended to other platforms depending on demand and drive.

Changes:
- Made skeleton .NET interface project (LIfebloom-GUI)
- Made skeleton c++ project (Lifebloom-Core)
- Moved old code to 'legacy'  until functionality is surpassed in Lifebloom-Core
- Implemented base classes for LSystem.h, FileReader.h and FileWriter.h
- Started to flesh out class methods of FileReader.h